Renombrando el archivo con el refactor y agregando más pruebas
